# contracts/agent.openapi.yaml
openapi: 3.1.0
info:
  title: Shadow Foliant Agent API
  version: 0.6.0
servers:
  - url: https://stock.example.com
security:
  - agentBearer: []
paths:
  /api/machine/v1/agent/research/cases:
    get:
      operationId: get_agent_research_cases
      summary: Read candidate research questions and public event reviews, without private theses
      responses:
        "200":
          description: Bounded research case list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ToolResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/machine/v1/agent/decision-loop:
    get:
      operationId: get_agent_decision_loop
      summary: Read immutable selection context, experimental evidence and model ledger, without private holdings
      responses:
        "200":
          description: Research decision loop
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ToolResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/machine/v1/agent/portfolio/action-plan:
    get:
      operationId: get_agent_account_action_plan
      summary: Preview account actions from server-loaded holdings and formal candidates; never places orders
      description: Requires delegated primary-portfolio access. Cash must be user-confirmed; omit unknown cash. Stale quotes and missing sellability fail closed.
      parameters:
        - name: available_cash
          in: query
          schema: { type: number, minimum: 0, maximum: 10000000000 }
        - name: allow_add
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Private preview with expiration, blockers and holdings watermark
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ToolResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/machine/v1/agent/portfolio/books:
    get:
      operationId: get_agent_account_books
      summary: Read actual securities-subaccount returns, distinct from signal and model returns
      parameters:
        - name: days
          in: query
          schema: { type: integer, minimum: 1, maximum: 365, default: 30 }
      responses:
        "200":
          description: Private account book with explicit accounting scope
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ToolResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    agentBearer:
      type: http
      scheme: bearer
      bearerFormat: Shadow Agent token for audience foliant
...
  schemas:
    Provenance:
      type: object
      additionalProperties: false
      required:
        - decision_at
        - market_as_of
        - financial_cutoff_at
        - universe_snapshot_id
        - input_manifest_id
        - policy_hash
        - code_revision
        - run_id
      properties:
        decision_at: { type: string, nullable: true }
        market_as_of: { type: string, nullable: true }
        financial_cutoff_at: { type: string, nullable: true }
        universe_snapshot_id: { type: string, nullable: true }
        input_manifest_id: { type: string, nullable: true }
        policy_hash: { type: string, nullable: true }
        code_revision: { type: string, nullable: true }
        run_id: { type: string, nullable: true }
    ToolResult:
      type: object
      additionalProperties: false
      required: [summary, resource_uri, status, provenance, warnings, data]
      properties:
        summary: { type: string, maxLength: 2000 }
        resource_uri: { type: string, pattern: "^shadow://foliant/" }
        status: { enum: [complete, degraded, missing, failed] }
        provenance: { $ref: "#/components/schemas/Provenance" }
        warnings:
          type: array
          items: { type: string, maxLength: 500 }
        data: {}
        model_payload: {}
        continuation: { type: object, additionalProperties: true }
        derived_analysis: {}
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [ok, error]
      properties:
        ok: { const: false }
        error:
          oneOf:
            - { type: string }
            - type: object
              additionalProperties: false
              required: [code, message]
              properties:
                code: { type: string }
                message: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid Agent Bearer
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: Audience or required scope is not authorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...